fix: Fix yaml, refine code according to comments
