+++ examples/embodiment/config/maniskill_ppo_openvlaoft.yaml
@@ -45,7 +45,7 @@ algorithm:
   n_eval_chunk_steps: 10
   num_group_envs: 128
   rollout_epoch: 1
-  reward_type: step_level
+  reward_type: action_level
   logprob_type: token_level
   entropy_type: token_level
 
@@ -96,7 +96,7 @@ rollout:
     queue_size: 0
   mode: "colocate"
   backend: "huggingface"
-  model_dir: "/path/to/openvla/openvla-7b-finetuned-libero-10/"
+  model_dir: "/path/to/model/Openvla-oft-SFT-libero10-trajall/"
   enable_offload: True
   pipeline_stage_num: 1
 
@@ -107,7 +107,7 @@ actor:
     queue_name: "replay_buffer"
     queue_size: 0
   training_backend: "fsdp"
-  checkpoint_load_path: "/path/to/openvla/openvla-7b-finetuned-libero-10/"
+  checkpoint_load_path: "/path/to/model/Openvla-oft-SFT-libero10-trajall/"
   checkpoint_save_path: "../results"
   micro_batch_size: 40
   global_batch_size: 640
@@ -151,7 +151,7 @@ actor:
 
   tokenizer:
     tokenizer_type: "HuggingFaceTokenizer"
-    tokenizer_model: "/path/to/openvla/openvla-7b-finetuned-libero-10/"
+    tokenizer_model: "/path/to/model/Openvla-oft-SFT-libero10-trajall/"
     use_fast: False
     trust_remote_code: True
     padding_side: "right"
