+++ rlinf/models/embodiment/openpi_action_model.py
@@ -12,9 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import logging
 import random
-from collections import OrderedDict
 from collections.abc import Sequence
 from dataclasses import dataclass, field
 from typing import Any, Dict, List, Literal, Tuple
@@ -22,15 +20,13 @@ from typing import Any, Dict, List, Literal, Tuple
 import jax
 import numpy as np
 import torch
-import torch.nn as nn
 from openpi import transforms as _transforms
 from openpi.models import model as _model
 from openpi.models.pi0_config import Pi0Config
 from openpi.models_pytorch.pi0_pytorch import PI0Pytorch, make_att_2d_masks
-
+from rlinf.models.embodiment.modules.explore_noise_net import ExploreNoiseNet
 from rlinf.models.embodiment.modules.value_head import ValueHead
 
-logger = logging.getLogger(__name__)
 
 
 @dataclass(frozen=True)
@@ -187,22 +183,7 @@ class OpenPi0ForRLActionPrediction(PI0Pytorch):
         compute_entropy: bool = False,
         compute_values: bool = False,
     ) -> Dict[str, Any]:
-        """
-        Unified forward function for Pi0 model that handles both prediction and logprob computation.
 
-        Args:
-            batch: Input batch dictionary containing observations (predict mode) or training data (compute_logprob mode)
-            mode: "predict" for action prediction, "compute_logprob" for logprob computation
-            sample_mode: Sampling mode for prediction ("ode" or "sde") - not used in current predict implementation
-            rollout_stage: Whether in rollout stage - not used in current predict implementation
-            output_lang_tokens: Whether to output language tokens in prediction mode
-            requires_grad: Whether to enable gradients (auto-detected if None)
-
-        Returns:
-            Dictionary containing:
-            - For predict mode: actions, chains, log_probs, values, (optionally lang_tokens, lang_masks)
-            - For compute_logprob mode: token_level_entropy, token_level_log_probs, action_level_logprobs
-        """
         chains = data["chains"]
         denoise_inds = data["denoise_inds"]
         # input transform
@@ -252,7 +233,7 @@ class OpenPi0ForRLActionPrediction(PI0Pytorch):
             "entropy": None,
         }
 
-    def _process_obs_from_env(self, env_processed_obs):
+    def input_processor(self, env_processed_obs):
         to_process_obs = {
             "observation/image": env_processed_obs["images"],
             "observation/wrist_image": env_processed_obs["wrist_images"],
@@ -271,7 +252,6 @@ class OpenPi0ForRLActionPrediction(PI0Pytorch):
                 ]
             elif torch.is_tensor(value):
                 processed_obs[key] = value.to(device=device).contiguous()
-            # todo: patch for openpi
             elif isinstance(value, dict):
                 for sub_key, sub_value in value.items():
                     processed_obs[key][sub_key] = sub_value.to(
@@ -282,7 +262,7 @@ class OpenPi0ForRLActionPrediction(PI0Pytorch):
     def predict_action_batch(
         self, env_obs, mode: Literal["train", "eval"] = "train", compute_values=True
     ) -> Tuple[np.ndarray, Dict[str, Any]]:
-        processed_obs = self._process_obs_from_env(env_obs)
+        processed_obs = self.input_processor(env_obs)
         observation = _model.Observation.from_dict(processed_obs)
         outputs = self.sample_actions(
             observation, mode=mode, compute_values=compute_values
@@ -670,151 +650,3 @@ class OpenPi0ForRLActionPrediction(PI0Pytorch):
             for params in self.paligemma_with_expert.paligemma.parameters():
                 params.requires_grad = False
 
-
-class ExploreNoiseNet(nn.Module):
-    """
-    Neural network to generate learnable exploration noise, conditioned on time embeddings and or state embeddings.
-    """
-
-    def __init__(
-        self,
-        in_dim: int,
-        out_dim: int,
-        hidden_dims: List[int],
-        activation_type: str,
-        noise_logvar_range: list,  # [min_std, max_std]
-        noise_scheduler_type: str,
-    ):
-        super().__init__()
-        self.mlp_logvar = MLP(
-            [in_dim] + hidden_dims + [out_dim],
-            activation_type=activation_type,
-            out_activation_type="Identity",
-        )
-        self.noise_scheduler_type = noise_scheduler_type
-        self.set_noise_range(noise_logvar_range)
-
-    def set_noise_range(self, noise_logvar_range: list):
-        self.noise_logvar_range = noise_logvar_range
-        noise_logvar_min = self.noise_logvar_range[0]
-        noise_logvar_max = self.noise_logvar_range[1]
-        self.register_buffer(
-            "logvar_min",
-            torch.log(torch.tensor(noise_logvar_min**2, dtype=torch.float32)).unsqueeze(
-                0
-            ),
-        )
-        self.register_buffer(
-            "logvar_max",
-            torch.log(torch.tensor(noise_logvar_max**2, dtype=torch.float32)).unsqueeze(
-                0
-            ),
-        )
-
-    def forward(self, noise_feature: torch.Tensor):
-        if "const" in self.noise_scheduler_type:  # const or const_schedule_itr
-            # pick the lowest noise level when we use constant noise schedulers.
-            noise_std = torch.exp(0.5 * self.logvar_min)
-        else:
-            # use learnable noise level.
-            noise_logvar = self.mlp_logvar(noise_feature)
-            noise_std = self.post_process(noise_logvar)
-        return noise_std
-
-    def post_process(self, noise_logvar):
-        """
-        input:
-            torch.Tensor([B, Ta , Da])
-        output:
-            torch.Tensor([B, Ta, Da])
-        """
-        noise_logvar = torch.tanh(noise_logvar)
-        noise_logvar = (
-            self.logvar_min
-            + (self.logvar_max - self.logvar_min) * (noise_logvar + 1) / 2.0
-        )
-        noise_std = torch.exp(0.5 * noise_logvar)
-        return noise_std
-
-
-activation_dict = nn.ModuleDict(
-    {
-        "relu": nn.ReLU(),
-        "elu": nn.ELU(),
-        "gelu": nn.GELU(),
-        "tanh": nn.Tanh(),
-        "mish": nn.Mish(),
-        "identity": nn.Identity(),
-        "softplus": nn.Softplus(),
-        "silu": nn.SiLU(),
-    }
-)
-
-
-class MLP(nn.Module):
-    def __init__(
-        self,
-        dim_list,
-        append_dim=0,
-        append_layers=None,
-        activation_type="tanh",
-        out_activation_type="identity",
-        use_layernorm=False,
-        use_layernorm_final=False,
-        dropout=0,
-        use_drop_final=False,
-        out_bias_init=None,
-        verbose=False,
-    ):
-        super(MLP, self).__init__()
-
-        # Ensure append_layers is always a list to avoid TypeError
-        self.append_layers = append_layers if append_layers is not None else []
-
-        # Construct module list
-        self.moduleList = nn.ModuleList()
-        num_layer = len(dim_list) - 1
-        for idx in range(num_layer):
-            i_dim = dim_list[idx]
-            o_dim = dim_list[idx + 1]
-            if append_dim > 0 and idx in self.append_layers:
-                i_dim += append_dim
-            linear_layer = nn.Linear(i_dim, o_dim)
-
-            # Add module components
-            layers = [("linear_1", linear_layer)]
-            if use_layernorm and (idx < num_layer - 1 or use_layernorm_final):
-                layers.append(("norm_1", nn.LayerNorm(o_dim)))  # type: ignore
-            if dropout > 0 and (idx < num_layer - 1 or use_drop_final):
-                layers.append(("dropout_1", nn.Dropout(dropout)))  # type: ignore
-
-            # Add activation function
-            act = (
-                activation_dict[activation_type.lower()]
-                if idx != num_layer - 1
-                else activation_dict[out_activation_type.lower()]
-            )
-            layers.append(("act_1", act))  # type: ignore
-
-            # Re-construct module
-            module = nn.Sequential(OrderedDict(layers))
-            self.moduleList.append(module)
-        if verbose:
-            logging.info(self.moduleList)
-
-        # Initialize the bias of the final linear layer if specified
-        if out_bias_init is not None:
-            final_linear = self.moduleList[-1][
-                0
-            ]  # Linear layer is first in the last Sequential # type: ignore
-            nn.init.constant_(final_linear.bias, out_bias_init)
-            logger.info(
-                f"Initialized the bias of the final linear layer to {out_bias_init}"
-            )
-
-    def forward(self, x, append=None):
-        for layer_ind, m in enumerate(self.moduleList):
-            if append is not None and layer_ind in self.append_layers:
-                x = torch.cat((x, append), dim=-1)
-            x = m(x)
-        return x
